+++ alf/algorithms/ppo_algorithm.py
@@ -38,12 +38,8 @@ class PPOAlgorithm(ActorCriticAlgorithm):
 
     def preprocess_experience(self, exp: Experience):
         """Compute advantages and put it into exp.info."""
-        reward = exp.reward
-        if self._reward_shaping_fn is not None:
-            reward = self._reward_shaping_fn(reward)
-        reward = self.calc_training_reward(reward, exp.info)
         advantages = value_ops.generalized_advantage_estimation(
-            rewards=reward,
+            rewards=exp.reward,
             values=exp.info.value,
             step_types=exp.step_type,
             discounts=exp.discount * self._loss._gamma,
