+++ alf/drivers/on_policy_driver.py
@@ -174,7 +174,7 @@ class OnPolicyDriver(policy_driver.PolicyDriver):
     def _train_loop_body(self, counter, time_step, policy_state,
                          training_info_ta):
 
-        next_time_step, policy_step, action = self._step(
+        next_time_step, policy_step, action, time_step = self._step(
             time_step, policy_state)
         action = tf.nest.map_structure(tf.stop_gradient, action)
         action_distribution_param = common.get_distribution_params(
@@ -227,12 +227,12 @@ class OnPolicyDriver(policy_driver.PolicyDriver):
                  name='iter_loop')
 
         if self._final_step_mode == OnPolicyDriver.FINAL_STEP_SKIP:
-            next_time_step, policy_step, action = self._step(
+            next_time_step, policy_step, action, time_step = self._step(
                 time_step, policy_state)
             next_state = policy_step.state
         else:
+            time_step = self._algorithm.transform_timestep(time_step)
             policy_step = common.algorithm_step(self._algorithm.rollout,
-                                                self._observation_transformer,
                                                 time_step, policy_state)
             action = common.sample_action_distribution(policy_step.action)
             next_time_step = time_step
