+++ alf/examples/ac_cart_pole.gin
@@ -7,7 +7,7 @@ create_environment.env_name="CartPole-v0"
 create_environment.num_parallel_environments=8
 
 # reward scaling
-ActorCriticAlgorithm.reward_shaping_fn = @reward_scaling
+Agent.reward_shaping_fn = @reward_scaling
 common.reward_scaling.scale = 0.01
 
 # algorithm config
@@ -23,12 +23,12 @@ value/ValueNetwork.fc_layer_params=(100,)
 
 ac/Adam.learning_rate=1e-3
 
-ActorCriticAlgorithm.action_spec=%action_spec
+Agent.action_spec=%action_spec
 ActorCriticAlgorithm.actor_network=@actor/ActorDistributionNetwork()
 ActorCriticAlgorithm.value_network=@value/ValueNetwork()
-ActorCriticAlgorithm.optimizer=@ac/Adam()
+Agent.optimizer=@ac/Adam()
 
-ActorCriticAlgorithm.gradient_clipping=10.0
+Agent.gradient_clipping=10.0
 ActorCriticLoss.entropy_regularization=1e-4
 ActorCriticLoss.gamma=0.98
 ActorCriticLoss.td_error_loss_fn=@element_wise_huber_loss
@@ -38,12 +38,10 @@ ActorCriticLoss.use_td_lambda_return=True
 # training config
 TrainerConfig.trainer=@on_policy_trainer
 TrainerConfig.unroll_length=8
-TrainerConfig.algorithm_ctor=@ActorCriticAlgorithm
+TrainerConfig.algorithm_ctor=@Agent
 TrainerConfig.num_iterations=20
 TrainerConfig.checkpoint_interval=3
 TrainerConfig.evaluate=False
 TrainerConfig.debug_summaries=True
 TrainerConfig.summarize_grads_and_vars=False
 TrainerConfig.summary_interval=5
-
-
