+++ alf/examples/icm_mountain_car.gin
@@ -31,21 +31,27 @@ ICMAlgorithm.feature_spec=@icm/TensorSpec()
 ICMAlgorithm.encoding_net=@icm/EncodingNetwork()
 ICMAlgorithm.hidden_size=(200, 200)
 
-ActorCriticAlgorithm.action_spec=%action_spec
+Agent.action_spec=%action_spec
 ActorCriticAlgorithm.actor_network=@actor/ActorDistributionNetwork()
 ActorCriticAlgorithm.value_network=@value/ValueNetwork()
-ActorCriticAlgorithm.optimizer=@ac/Adam()
-ActorCriticAlgorithm.intrinsic_curiosity_module=@ICMAlgorithm()
-ActorCriticLoss.entropy_regularization=0.1
+Agent.optimizer=@ac/Adam()
+Agent.intrinsic_curiosity_module=@ICMAlgorithm()
 ActorCriticLoss.use_gae=True
 ActorCriticLoss.use_td_lambda_return=True
+
+Agent.enforce_entropy_target=True
+EntropyTargetAlgorithm/Adam.learning_rate=1e-2
+EntropyTargetAlgorithm/Adam.beta_1=0.5
+EntropyTargetAlgorithm/Adam.beta_2=0.9
+EntropyTargetAlgorithm.min_alpha=0.1
+EntropyTargetAlgorithm.optimizer=@EntropyTargetAlgorithm/Adam()
 
 EncodingNetwork.activation_fn=@tf.nn.relu
 
 # training config
 TrainerConfig.trainer=@on_policy_trainer
 TrainerConfig.unroll_length=100
-TrainerConfig.algorithm_ctor=@ActorCriticAlgorithm
+TrainerConfig.algorithm_ctor=@Agent
 TrainerConfig.num_iterations=1000000
 TrainerConfig.use_tf_functions=1
 TrainerConfig.debug_summaries=1
