+++ alf/examples/icm_super_mario_intrinsic_only.gin
@@ -6,5 +6,5 @@ include 'icm_super_mario.gin'
 ac/Adam.learning_rate=3e-5
 
 # use only intrinsic reward for training policy
-ActorCriticAlgorithm.extrinsic_reward_coef=0.0
-ActorCriticAlgorithm.intrinsic_reward_coef=1.0
+Agent.extrinsic_reward_coef=0.0
+Agent.intrinsic_reward_coef=1.0
