+++ alf/examples/off_policy_ac_breakout.gin
@@ -49,12 +49,12 @@ value/ValueNetwork.conv_layer_params=%CONV_LAYER_PARAMS
 
 ac/Adam.learning_rate=1e-3
 
-ActorCriticAlgorithm.action_spec=%action_spec
+Agent.action_spec=%action_spec
 ActorCriticAlgorithm.actor_network=@actor/ActorDistributionNetwork()
 ActorCriticAlgorithm.value_network=@value/ValueNetwork()
-ActorCriticAlgorithm.optimizer=@ac/Adam()
-ActorCriticAlgorithm.gradient_clipping=None
-ActorCriticAlgorithm.reward_shaping_fn=@reward_clipping
+Agent.optimizer=@ac/Adam()
+Agent.gradient_clipping=None
+Agent.reward_shaping_fn=@reward_clipping
 common.reward_clipping.minmax=(-1, 1)
 
 # driver config
@@ -74,10 +74,10 @@ TrainerConfig.mini_batch_length=None
 TrainerConfig.mini_batch_size=None
 TrainerConfig.use_tf_functions=True
 TrainerConfig.num_envs = %N
-TrainerConfig.algorithm_ctor=@ActorCriticAlgorithm
+TrainerConfig.algorithm_ctor=@Agent
 TrainerConfig.debug_summaries = 1
 TrainerConfig.evaluate = False
 
-PolicyDriver.observation_transformer=@image_scale_transformer
+Agent.observation_transformer=@image_scale_transformer
 common.image_scale_transformer.min=0.0
