+++ alf/examples/ppo_async_icm_super_mario_intrinsic_only.gin
@@ -44,16 +44,16 @@ ICMAlgorithm.feature_spec=@icm/TensorSpec()
 ICMAlgorithm.encoding_net=@icm/EncodingNetwork()
 ICMAlgorithm.hidden_size=256
 
-ActorCriticAlgorithm.action_spec=%action_spec
+Agent.action_spec=%action_spec
 ActorCriticAlgorithm.actor_network=@actor/ActorDistributionNetwork()
 ActorCriticAlgorithm.value_network=@value/ValueNetwork()
-ActorCriticAlgorithm.optimizer=@ac/Adam()
-ActorCriticAlgorithm.intrinsic_curiosity_module=@ICMAlgorithm()
-ActorCriticAlgorithm.gradient_clipping=10.0
+Agent.optimizer=@ac/Adam()
+Agent.intrinsic_curiosity_module=@ICMAlgorithm()
+Agent.gradient_clipping=10.0
 
 # use only intrinsic reward for training policy
-ActorCriticAlgorithm.extrinsic_reward_coef=0.0
-ActorCriticAlgorithm.intrinsic_reward_coef=1.0
+Agent.extrinsic_reward_coef=0.0
+Agent.intrinsic_reward_coef=1.0
 
 PPOLoss.entropy_regularization=1e-2
 PPOLoss.gamma=0.99
@@ -85,4 +85,4 @@ TrainerConfig.use_rollout_state = True
 TrainerConfig.use_tf_functions = True
 TrainerConfig.num_envs = %N
 
-PolicyDriver.observation_transformer=@image_scale_transformer
+Agent.observation_transformer=@image_scale_transformer
