+++ alf/examples/ppo_bullet_humanoid.gin
@@ -11,8 +11,8 @@ create_environment.env_name="HumanoidBulletEnv-v0"
 create_environment.num_parallel_environments=96
 
 # algorithm config
-ActorCriticAlgorithm.gradient_clipping=0.5
-ActorCriticAlgorithm.clip_by_global_norm=True
+Agent.gradient_clipping=0.5
+Agent.clip_by_global_norm=True
 
 PPOLoss.entropy_regularization=1e-2
 PPOLoss.gamma=0.99
@@ -35,10 +35,10 @@ value/ValueNetwork.activation_fn=@tf.nn.tanh
 
 ac/Adam.learning_rate=3e-4
 
-ActorCriticAlgorithm.action_spec=%action_spec
+Agent.action_spec=%action_spec
 ActorCriticAlgorithm.actor_network=@actor/ActorDistributionNetwork()
 ActorCriticAlgorithm.value_network=@value/ValueNetwork()
-ActorCriticAlgorithm.optimizer=@ac/Adam()
+Agent.optimizer=@ac/Adam()
 
 
 # training config
