+++ alf/examples/ppo_cart_pole.gin
@@ -5,7 +5,7 @@ create_environment.env_name="CartPole-v0"
 create_environment.num_parallel_environments=8
 
 # reward scaling
-ActorCriticAlgorithm.reward_shaping_fn = @reward_scaling
+Agent.reward_shaping_fn = @reward_scaling
 common.reward_scaling.scale = 0.01
 
 # algorithm config
@@ -18,10 +18,10 @@ value/ValueNetwork.fc_layer_params=(100,)
 
 ac/Adam.learning_rate=1e-3
 
-ActorCriticAlgorithm.action_spec=%action_spec
+Agent.action_spec=%action_spec
 ActorCriticAlgorithm.actor_network=@actor/ActorDistributionNetwork()
 ActorCriticAlgorithm.value_network=@value/ValueNetwork()
-ActorCriticAlgorithm.optimizer=@ac/Adam()
+Agent.optimizer=@ac/Adam()
 
 
 PPOLoss.entropy_regularization=1e-4
