+++ alf/examples/rnd_super_mario.gin
@@ -44,14 +44,14 @@ rnd/EncodingNetwork.activation_fn=@tf.nn.elu
 RNDAlgorithm.target_net=@rnd/EncodingNetwork()
 RNDAlgorithm.predictor_net=@rnd/EncodingNetwork()
 
-ActorCriticAlgorithm.action_spec=%action_spec
+Agent.action_spec=%action_spec
 ActorCriticAlgorithm.actor_network=@actor/ActorDistributionNetwork()
 ActorCriticAlgorithm.value_network=@value/ValueNetwork()
-ActorCriticAlgorithm.optimizer=@ac/Adam()
-ActorCriticAlgorithm.intrinsic_curiosity_module=@RNDAlgorithm()
-ActorCriticAlgorithm.extrinsic_reward_coef=0.0
+Agent.optimizer=@ac/Adam()
+Agent.intrinsic_curiosity_module=@RNDAlgorithm()
+Agent.extrinsic_reward_coef=0.0
 ActorCriticLoss.entropy_regularization=0.01
-ActorCriticAlgorithm.gradient_clipping=10.0
+Agent.gradient_clipping=10.0
 ActorCriticLoss.use_gae=True
 ActorCriticLoss.use_td_lambda_return=True
 
@@ -59,13 +59,13 @@ tf.keras.layers.Conv2D.padding="same"
 
 # training config
 TrainerConfig.unroll_length=32
-TrainerConfig.algorithm_ctor=@ActorCriticAlgorithm
+TrainerConfig.algorithm_ctor=@Agent
 TrainerConfig.num_iterations=10000000
 TrainerConfig.use_tf_functions=True
 TrainerConfig.debug_summaries=True
 TrainerConfig.summarize_grads_and_vars=False
 TrainerConfig.summary_interval=10
 
-PolicyDriver.observation_transformer=@image_scale_transformer
+Agent.observation_transformer=@image_scale_transformer
 
 TrainerConfig.trainer=@on_policy_trainer
